requirements.txt
PyPDF2>=3.0.0
python-docx>=1.0.0
docx2txt>=0.8
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
jieba>=0.42.1
pypandoc>=1.11
tqdm>=4.65.0 